requirements.txt
absl-py==0.5.0
astor==0.7.1
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
certifi==2018.8.24
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
numpy==1.14.5
pbr==4.2.0
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
pylint==1.9.3
requests==2.19.1
requests-toolbelt==0.8.0
SimpleCV==1.3
singledispatch==3.4.0.3
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
tensorflow-hub==0.1.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
